Refactor Compatibility package
