+++ src/main/java/p76/bicycles/service/compatibility/CompatibilityResult.java
@@ -1,4 +1,4 @@
-package p76.bicycles.service;
+package p76.bicycles.service.compatibility;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
